我也不知道是什么bug，但是他就这么发生了

+++ src/New/Entity/AttachEffectClass.cpp
@@ -275,7 +275,7 @@ void AttachEffectClass::CreateAnim()
 
 	const CoordStruct& FLH = pType->Anim_FLH.isset()
 		? pType->Anim_FLH
-		: (this->AnimIndex < pType->Anim_FLHs.size()
+		: (this->AnimIndex < static_cast<int>(pType->Anim_FLHs.size())
 			? pType->Anim_FLHs[this->AnimIndex]
 			: CoordStruct::Empty);
